node/mempool_precheck: PR #1422 wave-11 split floor helpers to mempool_precheck_floor.go (Codacy file-budget pass 2)
Wave-11 closes the residual mempool_precheck.go file-cumulative
complexity warning from wave-10 by extracting the
cheapFeeFloorPrecheck + precheckEarlyDefer pair into a fourth
sibling file mempool_precheck_floor.go in the same package.

Final layout (4 files, each within per-file budget):
- mempool_precheck.go (orchestration): buildPolicyInputSnapshotIfNeeded,
  checkTxParseAndContext, checkTransactionWithSnapshot
- mempool_precheck_floor.go (floor-driven fast-reject):
  cheapFeeFloorPrecheck, precheckEarlyDefer
- mempool_precheck_input.go (input-side helpers):
  feePrecheckP2PKInputValue, precheckP2PKInputStructurallyValid,
  precheckCoinbaseImmature
- mempool_precheck_output.go (output-side helper):
  feePrecheckP2PKOutputValue

Same package, no API change, no test impact. Each file's cumulative
complexity now within the per-file budget.

Wave-11 cumulative gate verification:
- gofmt -l ./node clean
- go vet ./node/... clean
- go build ./node/... clean
- go test ./node -run TestMempool -count=1 PASS
